==> habitat_networks/03_habitat_diff_from_random.Rmd <==
---
title: "03_habitat_diff_from_random"
output: html_document
date: "2023-08-21"
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

```{r}
# otu tables
fung_data <- readRDS("~/cmaiki_lts/kaciekaj/waimea/intermediates/habitat/fung_downsampled_otu_tables_by_hab.rds")
bact_data <- readRDS("~/cmaiki_lts/kaciekaj/waimea/intermediates/habitat/bact_downsampled_otu_tables_by_hab.rds")

remove_dummy_row <- function(df) {
  new_df <- df[row.names(df)!="dummy",]
  return(new_df)
}


fung_data <- lapply(fung_data, remove_dummy_row)
bact_data <- lapply(bact_data, remove_dummy_row)



# import igraph objects
fung_igraphs <- readRDS("../intermediates/habitat/fung_habitat_igraphs.rds")
bact_igraphs <- readRDS("../intermediates/habitat/bact_habitat_igraphs.rds")
cross_igraphs <- readRDS("../intermediates/habitat/cross_habitat_igraphs.rds")

# metadata
fung_meta <- readRDS("../intermediates/global/fully_filtered_p20_fungal_otu_metadata_matched_up.rds")
bact_meta <- readRDS("../intermediates/global/fully_filtered_p20_bact_otu_metadata_matched_up.rds")

habs <- unique(fung_meta$habitat)
```

## Diff from random

```{r}
fung_power_law <- lapply(fung_igraphs, function(x) power.law.fit(degree(x))$KS.p)
fung_power_law > 0.05

bact_power_law <- lapply(bact_igraphs, function(x) power.law.fit(degree(x))$KS.p)
bact_power_law > 0.05

cross_power_law <- lapply(cross_igraphs, function(x) power.law.fit(degree(x))$KS.p)
cross_power_law > 0.05


tab <- data.frame("Fungi" = unlist(fung_power_law),
                  "Bacteria" = unlist(bact_power_law),
                  "Cross-domain" = unlist(cross_power_law))
rownames(tab) <- habs

tab <- tab[c("Marine", "Riverine", "Terrestrial"),]

write.csv(tab, "../outputs/habitat_scalefree_table.csv")


# small world
library(qgraph)

fung_sw = lapply(fung_igraphs, smallworldIndex)
bact_sw <- lapply(bact_igraphs, smallworldIndex)
cross_sw <- lapply(cross_igraphs, smallworldIndex)

sw_vals <- data.frame(Fungi = sapply(fung_sw, function(x) x$index),
                      Bacteria = sapply(bact_sw, function(x) x$index),
                      Cross = sapply(cross_sw, function(x) x$index))

rownames(sw_vals) <- habs

sw_vals <- sw_vals[c("Marine", "Riverine", "Terrestrial"),]

write.csv(sw_vals, "../outputs/habitat_smallworld_table.csv")

```


### Independence btwn abundance and degree
```{r}
# regressions

### functions
# add reads column to degree df
add_reads <- function (deg_df, otu_table) {
  deg_df$reads <- rowSums(otu_table)
  return(deg_df)
}

#define function to extract overall p-value of model
overall_p <- function(my_model) {
    f <- summary(my_model)$fstatistic
    p <- pf(f[1],f[2],f[3],lower.tail=F)
    attributes(p) <- NULL
    return(p)
}


### fungi
fung_deg <- lapply(fung_igraphs, function(x) data.frame(degree = degree(x)))

all(names(fung_deg[[1]]) == rownames(fung_data[[1]]))

fung_dfs <- mapply(add_reads,
               fung_deg,
               fung_data,
               SIMPLIFY = FALSE)

fung_lms <- lapply(fung_dfs, function(x) lm(log(reads)~log(degree+1), data = x))

fung_r2 <- sapply(fung_lms, function(x) round(summary(x)$r.squared, 4))
fung_p <- sapply(fung_lms, function(x) round(overall_p(x), 4))

fung_reg_res <- data.frame(type = rep("Fungi", 3),
                           habitat = habs,
                           r2 = fung_r2,
                           p = fung_p)

fung_reg_res$habitat <- factor(fung_reg_res$habitat, levels=c("Marine", "Terrestrial", "Riverine"))
fung_reg_res <- fung_reg_res[order(fung_reg_res$habitat),]


### bacteria
bact_deg <- lapply(bact_igraphs, function(x) data.frame(degree = degree(x)))

all(rownames(bact_deg[[1]]) == rownames(bact_data[[1]]))

bact_dfs <- mapply(add_reads,
               bact_deg,
               bact_data,
               SIMPLIFY = FALSE)

bact_lms <- lapply(bact_dfs, function(x) lm(log(reads)~log(degree), data = x))

bact_r2 <- sapply(bact_lms, function(x) round(summary(x)$r.squared, 4))
bact_p <- sapply(bact_lms, function(x) round(overall_p(x), 4))

bact_reg_res <- data.frame(type = rep("Bacteria", 3),
                           habitat = habs,
                           r2 = bact_r2,
                           p = bact_p)

bact_reg_res$habitat <- factor(bact_reg_res$habitat, levels=c("Marine", "Terrestrial", "Riverine"))
bact_reg_res <- bact_reg_res[order(bact_reg_res$habitat),]

# pl <- ggplot(bact_dfs[[1]], aes(x=log(degree), y = log(reads))) +
#   geom_point() +
#   #ggtitle("Cross-domain") +
#   #geom_smooth(method = "lm", se = FALSE) +
#   theme(panel.border = element_blank(), panel.grid.major = element_blank(),
#         panel.grid.minor = element_blank(), axis.line = element_line(colour = "black"),
#         panel.background=element_blank()) +
#   labs(subtitle = paste("R2 =", round(summary(bact_lms[[1]])$r.squared, 3), ";", "P < 0.001")) +
#   scale_y_continuous(breaks = seq(4, 16, 4), limits = c(c(2.9, 16))) +
#   scale_x_continuous(breaks = seq(1, 7, 2), limits = c(c(1, 7)))


### cross
cross_deg <- lapply(cross_igraphs, function(x) data.frame(degree = degree(x)))

all(names(fung_data[[1]])==names(bact_data[[1]]))

cross_data <- mapply(rbind, fung_data, bact_data, SIMPLIFY = FALSE)

all(rownames(cross_deg[[1]]) == rownames(cross_data[[1]]))

cross_dfs <- mapply(add_reads,
               cross_deg,
               cross_data,
               SIMPLIFY = FALSE)

cross_lms <- lapply(cross_dfs, function(x) lm(log(reads)~log(degree), data = x))

cross_r2 <- sapply(cross_lms, function(x) round(summary(x)$r.squared, 4))
cross_p <- sapply(cross_lms, function(x) round(overall_p(x), 4))

cross_reg_res <- data.frame(type = rep("Cross", 3),
                           habitat = habs,
                           r2 = cross_r2,
                           p = cross_p)

cross_reg_res$habitat <- factor(cross_reg_res$habitat, levels=c("Marine", "Terrestrial", "Riverine"))
cross_reg_res <- cross_reg_res[order(cross_reg_res$habitat),]


all_data <- do.call("rbind", list(fung_reg_res, bact_reg_res, cross_reg_res))
names(all_data) <- c("Network Type", "Plot", "R2", "P")

write.csv(all_data, "../outputs/habitat_abun_vs_degree_regression_values.csv")

```
