==> habitat_networks/03a_habitat_weights.Rmd <==
---
title: "02_habitat_netviz"
output: html_document
date: "2023-06-19"
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

```{r, message=FALSE}
library(SpiecEasi)
library(igraph)
library(ggnetwork)
library(intergraph)
library(ggpubr)

options(bitmapType='cairo')

```

Get metadata and otu names for each plot network

```{r}
fung_data <- readRDS("../intermediates/habitat/fung_downsampled_otu_tables_by_hab.rds")
bact_data <- readRDS("../intermediates/habitat/bact_downsampled_otu_tables_by_hab.rds")

fung_meta <- readRDS("../intermediates/global/fully_filtered_p20_fungal_otu_metadata_matched_up.rds")
bact_meta <- readRDS("../intermediates/global/fully_filtered_p20_bact_otu_metadata_matched_up.rds")

habs <- unique(fung_meta$habitat)

```

## ITS

```{r}

fung_habitat_networks <- readRDS("../networks/prev20/habitat/outputs/fung_habitat_networks.rds")

fung_vertex_names <- lapply(fung_data, rownames)

lapply(fung_habitat_networks, function(x) sum(getRefit(x))/2)

# getOptInd(se)
# 
# sum(getRefit(se))/2

# add weights
weighted_fung_igraphs <- list()

for (i in 1:length(fung_habitat_networks)) {
  optb <- as.matrix(symBeta(getOptBeta(fung_habitat_networks[[i]])))
  edge_cols <- ifelse(optb>0, 'positive', 'negative')[upper.tri(optb) & optb!=0]
  ig <- adj2igraph(getRefit(fung_habitat_networks[[i]]),
                   vertex.attr=list(name=fung_vertex_names[[i]]),
                   edge.attr=list(edge_sign= edge_cols))
  
  weighted_fung_igraphs[[i]] <- ig
}

lapply(weighted_fung_igraphs, function(x) table(E(x)$edge_sign))


fung_no_dummy <- lapply(weighted_fung_igraphs, function(x) delete_vertices(x, "dummy"))

fung_weights <- lapply(fung_no_dummy, function(x) table(E(x)$edge_sign))

fung_wt_tbl = as.data.frame(do.call("rbind", fung_weights))
fung_wt_tbl$habitat <- habs

make_pct_tbl <- function(weight_table) {
  weight_table$total <- weight_table$negative + weight_table$positive
  weight_table$neg_pct <- weight_table$negative / weight_table$total
  weight_table$pos_pct <- weight_table$positive / weight_table$total
  
  #eight_table <- weight_table[c("Beach", "Estuary", "Entrance", "Confluence", "Waterfall", "DrumRoad", "Ridge"),]
  
  return(weight_table)
}

fung_wt_tbl <- make_pct_tbl(fung_wt_tbl)
fung_wt_tbl$domain <- "Fungi"




## visualization
# 
# habs <- unique(fung_meta$habitat)
# 
# 
# fung_net_plots <- list()
# 
# for (i in 1:length(fung_no_dummy)) {
#   set.seed(123)
#   
#   gg <- ggplot(ggnetwork(asNetwork(fung_no_dummy[[i]]), layout = "fruchtermanreingold"), aes(x = x, y = y, xend = xend, yend = yend)) +
#   geom_edges(color = "grey20", linewidth = 0.1) +
#   geom_nodes(color = "#1F78B4") +
#   theme_blank() +
#   #scale_color_brewer(palette = "Paired") + # "#A6CEE3" "#1F78B4"
#   labs(title = paste("Habitat:", habs[i]))
#   
#   fung_net_plots[[i]] <- gg
# }
# 
# names(fung_net_plots) <- habs
# fung_net_plots <- fung_net_plots[c("Marine", "Terrestrial", "Riverine")]
# 
# all_fung <- ggarrange(plotlist = fung_net_plots,
#                       ncol = 3,
#                       nrow = 1) + bgcolor("white")
# 
# ggsave("../figures/habitat/fung_network_plots.pdf", all_fung, width=18.5, height=6)

```

## 16S

```{r}

bact_habitat_networks <- readRDS("../networks/prev20/habitat/outputs/bact_habitat_networks.rds")

bact_vertex_names <- lapply(bact_data, rownames)

lapply(bact_habitat_networks, function(x) sum(getRefit(x))/2)

# getOptInd(se)
# 
# sum(getRefit(se))/2

# add weights
weighted_bact_igraphs <- list()

for (i in 1:length(bact_habitat_networks)) {
  optb <- as.matrix(symBeta(getOptBeta(bact_habitat_networks[[i]])))
  edge_cols <- ifelse(optb>0, 'positive', 'negative')[upper.tri(optb) & optb!=0]
  ig <- adj2igraph(getRefit(bact_habitat_networks[[i]]),
                   vertex.attr=list(name=bact_vertex_names[[i]]),
                   edge.attr=list(edge_sign= edge_cols))
  
  weighted_bact_igraphs[[i]] <- ig
}

lapply(weighted_bact_igraphs, function(x) table(E(x)$edge_sign))


bact_no_dummy <- lapply(weighted_bact_igraphs, function(x) delete_vertices(x, "dummy"))

bact_weights <- lapply(bact_no_dummy, function(x) table(E(x)$edge_sign))

bact_wt_tbl = as.data.frame(do.call("rbind", bact_weights))
bact_wt_tbl$habitat <- habs

bact_wt_tbl <- make_pct_tbl(bact_wt_tbl)
bact_wt_tbl$domain <- "Bacteria"

# 
# 
# ## visualization
# 
# habs <- unique(bact_meta$habitat)
# 
# 
# bact_net_plots <- list()
# 
# for (i in 1:length(bact_no_dummy)) {
#   set.seed(123)
#   
#   gg <- ggplot(ggnetwork(asNetwork(bact_no_dummy[[i]]), layout = "fruchtermanreingold"), aes(x = x, y = y, xend = xend, yend = yend)) +
#   geom_edges(color = "grey20", linewidth = 0.1) +
#   geom_nodes(color = "#A6CEE3") +
#   theme_blank() +
#   #scale_color_brewer(palette = "Paired") + # "#A6CEE3" "#1F78B4"
#   labs(title = paste("Habitat:", habs[i]))
#   
#   bact_net_plots[[i]] <- gg
# }
# 
# 
# names(bact_net_plots) <- habs
# bact_net_plots <- bact_net_plots[c("Marine", "Terrestrial", "Riverine")]
# 
# all_bact <- ggarrange(plotlist = bact_net_plots,
#                       ncol = 3,
#                       nrow = 1) + bgcolor("white")
# 
# ggsave("../figures/habitat/bact_network_plots.pdf", all_bact, width=18.5, height=6)

```

## Cross-domain

```{r}

cross_habitat_networks <- readRDS("../networks/prev20/habitat/outputs/cross_habitat_networks.rds")

#readRDS("../networks/prev20/habitat/outputs/cross_habitat_networks.rds")

cross_vertex_names <- list()

for (i in 1:3) {
  cross_names <- c(fung_vertex_names[[i]], bact_vertex_names[[i]])
  cross_vertex_names[[i]] <- cross_names
}

lapply(cross_habitat_networks, function(x) sum(getRefit(x))/2)

# getOptInd(se)
# 
# sum(getRefit(se))/2

# add weights
weighted_cross_igraphs <- list()

for (i in 1:length(cross_habitat_networks)) {
  optb <- as.matrix(symBeta(getOptBeta(cross_habitat_networks[[i]])))
  edge_cols <- ifelse(optb>0, 'positive', 'negative')[upper.tri(optb) & optb!=0]
  ig <- adj2igraph(getRefit(cross_habitat_networks[[i]]),
                   vertex.attr=list(name=cross_vertex_names[[i]]),
                   edge.attr=list(edge_sign= edge_cols))
  
  weighted_cross_igraphs[[i]] <- ig
}

lapply(weighted_cross_igraphs, function(x) table(E(x)$edge_sign))


cross_no_dummy <- lapply(weighted_cross_igraphs, function(x) delete_vertices(x, "dummy"))
cross_no_dummy <- lapply(cross_no_dummy, function(x) delete_vertices(x, "dummy"))

cross_weights <- lapply(cross_no_dummy, function(x) table(E(x)$edge_sign))

cross_wt_tbl = as.data.frame(do.call("rbind", cross_weights))
cross_wt_tbl$habitat <- habs

cross_wt_tbl <- make_pct_tbl(cross_wt_tbl)
cross_wt_tbl$domain <- "Interkingdom"


## visualization
# cross_net_plots <- list()
# 
# for (i in 1:length(cross_no_dummy)) {
#   set.seed(123)
#   
#   dat <- ggnetwork(asNetwork(cross_no_dummy[[i]]), layout = "fruchtermanreingold")
#   dat$domain <- ifelse(grepl("fung", dat$vertex.names), "Fungi", "Bacteria")
# 
#   
#   gg <- ggplot(dat, aes(x = x, y = y, xend = xend, yend = yend)) +
#   geom_edges(color = "grey20", linewidth = 0.1) +
#   geom_nodes(aes(color = domain)) +
#   theme_blank() +
#   scale_color_brewer(palette = "Paired") + # "#A6CEE3" "#1F78B4" +
#     guides(color = guide_legend(override.aes = list(size=5))) +
#   labs(title = paste("Habitat:", habs[i]), color = "Domain")
#   
#   cross_net_plots[[i]] <- gg
# }
# 
# 
# names(cross_net_plots) <- habs
# cross_net_plots <- cross_net_plots[c("Marine", "Terrestrial", "Riverine")]
# 
# all_cross <- ggarrange(plotlist = cross_net_plots,
#                       ncol = 3,
#                       nrow = 1,
#                       common.legend = TRUE,
#                       legend = "bottom") + bgcolor("white")
# 
# ggsave("../figures/habitat/cross_network_plots.pdf",all_cross, width=21, height=6)

```

Plot!
```{r}

all_dat <- do.call("rbind", list(fung_wt_tbl, bact_wt_tbl, cross_wt_tbl))
all_dat$domain <- factor(all_dat$domain, c("Fungi", "Bacteria", "Interkingdom"))
domains <- unique(all_dat$domain)

saveRDS(all_dat, "../intermediates/habitat/hab_neg_pct.rds")

neg_edge_plots <- list()

my_colors <- wes_palette("GrandBudapest2")[2:4]

  pl <- ggplot(all_dat, aes(x = habitat, y = neg_pct, fill = domain)) +
    geom_bar(stat="identity", position="dodge") +
    scale_fill_manual(values = my_colors) +
    theme_classic() +
    theme(axis.title.y=element_text(colour="black", margin=margin(0,5,0,0))) +
    theme(axis.title.x=element_text(colour="black",margin=margin(5,0,0,0))) +
    theme(axis.text.x=element_text(colour="black",size=10)) +
    theme(text=element_text(colour="black",size=12)) +
    theme(axis.text.y=element_text(colour="black",size=9)) +
    labs(fill = "Kingdom", x = "Habitat", y = "% Negative Edges") +
    scale_y_continuous(breaks = seq(0, 0.35, 0.05), limits = c(c(0, 0.35)))
  
ggsave("../figures/finals/habitat/neg_edges_by_habitat.pdf", width=8, height=4.5)

```


```{r}
saveRDS(list(weighted_fung_igraphs, weighted_bact_igraphs, weighted_cross_igraphs),
        "../intermediates/habitat/weighted_hab_igraphs.rds")

```

## TEST on how much of interkingdom edges are pos/neg
```{r}
weighted_cross_igraphs <- readRDS("../intermediates/habitat/weighted_hab_igraphs.rds")[[3]]

edgelists = lapply(weighted_cross_igraphs, function(x) as_data_frame(x)[, ])

fung_bact <- lapply(edgelists, function(x) x[substr(x$from, 1, 3)!=substr(x$to, 1, 3),])

fung_only <- lapply(edgelists, function(x) x[grepl("fung", x$from) & grepl("fung", x$to),])

bact_only <- lapply(edgelists, function(x) x[grepl("bact", x$from) & grepl("bact", x$to),])

out = data.frame(rbind(table(fung_only$sign), table(bact_only$sign), table(fung_bact$sign)))
rownames(out) <- c("Fungi", "Bacteria", "Both")

out$neg_pct <- out$negative/(out$negative + out$positive)

try <- list(fung_only, bact_only, fung_bact)

hm <- lapply(try, function(y) lapply(y, function(x) table(x$edge_sign)))

wha <- lapply(hm, function(x) do.call("rbind", x))
wha2 <- do.call("rbind", wha)

final <- as.data.frame(wha2)

final$kingdom <- c(rep("Fungi", 3), rep("Bacteria", 3), rep("Both", 3))
final$habitat <- rep(habs, 3)

final$neg_pct <- final$negative/(final$negative + final$positive)

saveRDS(final, "../intermediates/habitat/table_of_interkingdom_edges_habitat.rds")


```
