==> habitat_networks/04_habitat_metrics.Rmd <==
---
title: "03_habitat_metrics"
output: html_document
date: "2023-06-19"
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

```{r, message=FALSE}
library(igraph)
library(ggnetwork)
library(network)
library(intergraph)

library(dplyr)
library(tibble)
library(pbmcapply)
library(reshape2)
library(wesanderson)
library(scales)
```

```{r}
options(bitmapType='cairo')
```

```{r}
# import igraph objects
fung_igraphs <- readRDS("../intermediates/habitat/fung_habitat_igraphs.rds")
bact_igraphs <- readRDS("../intermediates/habitat/bact_habitat_igraphs.rds")
cross_igraphs <- readRDS("../intermediates/habitat/cross_habitat_igraphs.rds")

# metadata
fung_meta <- readRDS("../intermediates/global/fully_filtered_p20_fungal_otu_metadata_matched_up.rds")
bact_meta <- readRDS("../intermediates/global/fully_filtered_p20_bact_otu_metadata_matched_up.rds")

habs <- unique(fung_meta$habitat)
```

### Metrics to Calculate

## ITS

```{r}
fung_nets <- fung_igraphs


## BETWEENNESS CENTRALITY
fung_bc <- sapply(fung_nets, betweenness)

## DEGREE
fung_deg <- sapply(fung_nets, degree)

## CONNECTANCE
connectance <- function(g) {
  ecount(g) / vcount(g)^2 
}


fung_conn <- sapply(fung_nets, connectance)


## EDGE TO NODE RATIO
fung_num_nodes <- sapply(fung_nets, gorder)
fung_num_edges <- sapply(fung_nets, gsize)

fung_eton_ratio <- fung_num_edges / fung_num_nodes

## MODULARITY
fung_netcarto <- readRDS("../intermediates/habitat/fung_habitat_netcarto_results.rds")

fung_new_mod <- sapply(fung_netcarto, function(x) x[[2]])



metrics <- c("Modularity", "Connectance", "Edge to Node Ratio")

fung_metrics <- data.frame(modularity <- fung_new_mod,
                           connectance <- fung_conn,
                           eton_ratio <- fung_eton_ratio
                           )
names(fung_metrics) <- metrics
fung_metrics$habitat <- habs

fung_mt_long <- melt(fung_metrics, id.vars = "habitat", variable.name = "category", value.name = "value")
names(fung_mt_long) <- c("habitat", "metric", "value")
fung_mt_long$type <- "Fungi"


unadjusted_fung_metric_plots <- list()

for (i in 1:length(metrics)) {
  dat <- fung_mt_long[fung_mt_long$metric==metrics[i],]
  #dat$habitat <- factor(dat$habitat, levels=c("Riverine", "Terrestrial", "Marine"))
  
  pl <- ggplot(dat, aes(x = habitat, y = value)) +
    geom_bar(stat="identity", fill=wes_palette("GrandBudapest2")[2]) +
    theme_classic() +
    theme(axis.text.x=element_text(colour="black",size=9)) +
    theme(text=element_text(colour="black",size=12)) +
    theme(axis.text.y=element_text(colour="black",size=9)) +
    theme(axis.title.y=element_blank()) +
    theme(axis.title.x=element_blank()) +
    labs(title = paste("ITS,", metrics[i]))
  
  unadjusted_fung_metric_plots[[i]] <- pl
}


## node metric boxplots
fung_deg_dat <- list()
for (i in 1:3) {
  df <- data.frame(habitat = rep(habs[[i]], length(fung_deg[[i]])),
                   metric = rep("degree", length(fung_deg[[i]])),
                   value = fung_deg[[i]])
  
  fung_deg_dat[[i]] <- df
}

fung_deg_all <- do.call("rbind", fung_deg_dat)
fung_deg_all$log_val <- log(fung_deg_all$value + 1)
fung_deg_all$type <- "Fungi"


fung_bc_dat <- list()
for (i in 1:3) {
  df <- data.frame(habitat = rep(habs[[i]], length(fung_bc[[i]])),
                   metric = rep("betweenness", length(fung_bc[[i]])),
                   value = fung_bc[[i]])

  fung_bc_dat[[i]] <- df
}

fung_bc_all <- do.call("rbind", fung_bc_dat)
fung_bc_all$log_val <- log(fung_bc_all$value + 1)
fung_bc_all$type <- "Fungi"


library(wesanderson)

wes_palette("GrandBudapest2")

fung_deg_boxplot <- ggplot(fung_deg_all, aes(x=habitat, y=log_val)) +
    geom_jitter(color=wes_palette("GrandBudapest2")[2], size=0.4, alpha=0.4) +
    geom_boxplot(fill = wes_palette("GrandBudapest2")[2]) +
    theme_classic() +
    theme(axis.text.x=element_text(colour="black",size=9)) +
    theme(text=element_text(colour="black",size=12)) +
    theme(axis.text.y=element_text(colour="black",size=9)) +
    theme(axis.title.y=element_blank()) +
    theme(axis.title.x=element_blank()) +
    labs(title = "ITS, Degree, log-transformed")


fung_bc_boxplot <- ggplot(fung_bc_all, aes(x=habitat, y=log_val)) +
    geom_jitter(color=wes_palette("GrandBudapest2")[1], size=0.4, alpha=0.4) +
    geom_boxplot(fill = wes_palette("GrandBudapest2")[1]) +
    theme_classic() +
    scale_y_continuous(breaks = seq(0, 10, 2.5), limits = c(c(0, 10))) + # throws warning but looks same as when commented out?
    theme(axis.text.x=element_text(colour="black",size=9)) +
    theme(text=element_text(colour="black",size=12)) +
    theme(axis.text.y=element_text(colour="black",size=9)) +
    theme(axis.title.y=element_blank()) +
    theme(axis.title.x=element_blank()) +
    labs(title = "ITS, Betweenness Centrality, log-transformed")




fung_metric_plots <- unadjusted_fung_metric_plots


# modularity
fung_metric_plots[[1]] <- fung_metric_plots[[1]] +
                          scale_y_continuous(breaks = seq(0, 0.8, 0.2),
                                             limits = c(c(0, 0.8)))

# diameter
fung_metric_plots[[2]] <- unadjusted_fung_metric_plots[[2]] +
                          scale_y_continuous(breaks = seq(0, 30, 6),
                                             limits = c(c(0, 30)))


# transitivity
fung_metric_plots[[3]] <- unadjusted_fung_metric_plots[[3]] +
                          scale_y_continuous(breaks = seq(0, 0.6, 0.2),
                                             limits = c(c(0, 0.6)))

# complexity
fung_metric_plots[[4]] <- fung_metric_plots[[4]] +
                          scale_y_continuous(breaks = seq(0, 0.006, 0.002),
                                             limits = c(c(0, 0.006)))



all_fung <- ggarrange(fung_deg_boxplot,
                      fung_bc_boxplot,
                      fung_metric_plots[[1]],
                      fung_metric_plots[[2]],
                      fung_metric_plots[[3]],
                      fung_metric_plots[[4]],
                      ncol = 3,
                      nrow = 2) + bgcolor("white")

#ggsave("../figures/habitat/fung_metric_plots.jpg", width=19, height=9.5)

```


## 16S

```{r}
bact_nets <- bact_igraphs


## BETWEENNESS CENTRALITY
bact_bc <- sapply(bact_nets, betweenness)

## DEGREE
bact_deg <- sapply(bact_nets, degree)

## CONNECTANCE
connectance <- function(g) {
  ecount(g) / vcount(g)^2 
}


bact_conn <- sapply(bact_nets, connectance)


## EDGE TO NODE RATIO
bact_num_nodes <- sapply(bact_nets, gorder)
bact_num_edges <- sapply(bact_nets, gsize)

bact_eton_ratio <- bact_num_edges / bact_num_nodes

## MODULARITY
bact_netcarto <- readRDS("../intermediates/habitat/bact_habitat_netcarto_results.rds")

bact_new_mod <- sapply(bact_netcarto, function(x) x[[2]])



metrics <- c("Modularity", "Connectance", "Edge to Node Ratio")

bact_metrics <- data.frame(modularity <- bact_new_mod,
                           connectance <- bact_conn,
                           eton_ratio <- bact_eton_ratio
                           )
names(bact_metrics) <- metrics
bact_metrics$habitat <- habs

bact_mt_long <- melt(bact_metrics, id.vars = "habitat", variable.name = "category", value.name = "value")
names(bact_mt_long) <- c("habitat", "metric", "value")
bact_mt_long$type <- "Bacteria"


## node metric boxplots
bact_deg_dat <- list()
for (i in 1:3) {
  df <- data.frame(habitat = rep(habs[[i]], length(bact_deg[[i]])),
                   metric = rep("degree", length(bact_deg[[i]])),
                   value = bact_deg[[i]])
  
  bact_deg_dat[[i]] <- df
}

bact_deg_all <- do.call("rbind", bact_deg_dat)
bact_deg_all$log_val <- log(bact_deg_all$value + 1)
bact_deg_all$type <- "Bacteria"


bact_bc_dat <- list()
for (i in 1:3) {
  df <- data.frame(habitat = rep(habs[[i]], length(bact_bc[[i]])),
                   metric = rep("betweenness", length(bact_bc[[i]])),
                   value = bact_bc[[i]])

  bact_bc_dat[[i]] <- df
}

bact_bc_all <- do.call("rbind", bact_bc_dat)
bact_bc_all$log_val <- log(bact_bc_all$value + 1)
bact_bc_all$type <- "Bacteria"

```


## Cross

```{r}
cross_nets <- cross_igraphs


## BETWEENNESS CENTRALITY
cross_bc <- sapply(cross_nets, betweenness)

## DEGREE
cross_deg <- sapply(cross_nets, degree)

## CONNECTANCE
connectance <- function(g) {
  ecount(g) / vcount(g)^2 
}


cross_conn <- sapply(cross_nets, connectance)


## EDGE TO NODE RATIO
cross_num_nodes <- sapply(cross_nets, gorder)
cross_num_edges <- sapply(cross_nets, gsize)

cross_eton_ratio <- cross_num_edges / cross_num_nodes

## MODULARITY
cross_netcarto <- readRDS("../intermediates/habitat/cross_habitat_netcarto_results.rds")

cross_new_mod <- sapply(cross_netcarto, function(x) x[[2]])



metrics <- c("Modularity", "Connectance", "Edge to Node Ratio")

cross_metrics <- data.frame(modularity <- cross_new_mod,
                           connectance <- cross_conn,
                           eton_ratio <- cross_eton_ratio
                           )
names(cross_metrics) <- metrics
cross_metrics$habitat <- habs

cross_mt_long <- melt(cross_metrics, id.vars = "habitat", variable.name = "category", value.name = "value")
names(cross_mt_long) <- c("habitat", "metric", "value")
cross_mt_long$type <- "Interkingdom"


## node metric boxplots
cross_deg_dat <- list()
for (i in 1:3) {
  df <- data.frame(habitat = rep(habs[[i]], length(cross_deg[[i]])),
                   metric = rep("degree", length(cross_deg[[i]])),
                   value = cross_deg[[i]])
  
  cross_deg_dat[[i]] <- df
}

cross_deg_all <- do.call("rbind", cross_deg_dat)
cross_deg_all$log_val <- log(cross_deg_all$value + 1)
cross_deg_all$type <- "Interkingdom"


cross_bc_dat <- list()
for (i in 1:3) {
  df <- data.frame(habitat = rep(habs[[i]], length(cross_bc[[i]])),
                   metric = rep("betweenness", length(cross_bc[[i]])),
                   value = cross_bc[[i]])

  cross_bc_dat[[i]] <- df
}

cross_bc_all <- do.call("rbind", cross_bc_dat)
cross_bc_all$log_val <- log(cross_bc_all$value + 1)
cross_bc_all$type <- "Interkingdom"

```


# Combine bars

```{r}
all_metric_data <- do.call("rbind", list(fung_mt_long, bact_mt_long, cross_mt_long))
all_metric_data$value <- as.numeric(all_metric_data$value)

all_metric_data$type <- factor(all_metric_data$type, levels = c("Fungi", "Bacteria", "Interkingdom"))


# consistent guides
g <- guides(fill = guide_legend(override.aes = list(fill =  paletteer_d("vapoRwave::floralShoppe")[2:4],
                                                    color = "black")))

unadjusted_full_metric_plots <- list()

for (i in 1:length(metrics)) {
  dat <- all_metric_data[all_metric_data$metric==metrics[i],]
  
  pl <- ggplot(dat, aes(x = habitat, y = value, fill = type)) +
    geom_bar(stat="identity", position="dodge", key_glyph = "rect") +
    scale_fill_manual(values =  paletteer_d("vapoRwave::floralShoppe")[2:4]) +
    theme_classic() +
    theme(axis.text.x=element_text(colour="black",size=9)) +
    theme(text=element_text(colour="black",size=12)) +
    theme(axis.text.y=element_text(colour="black",size=9)) +
    theme(axis.title.y=element_blank()) +
    theme(axis.title.x=element_blank()) +
    labs(title = metrics[i], fill = "Kingdom") +
    g
  
  unadjusted_full_metric_plots[[i]] <- pl
}

full_metric_plots <- unadjusted_full_metric_plots

# modularity
full_metric_plots[[1]] <- unadjusted_full_metric_plots[[1]] +
                          scale_y_continuous(breaks = seq(0, 0.7, 0.1),
                                             limits = c(c(0, 0.7)))


full_metric_plots[[3]] <- unadjusted_full_metric_plots[[3]] +
                          scale_y_continuous(breaks = seq(0, 25, 5),
                                             limits = c(c(0, 25)))



## degree boxplots
all_deg_dat <- do.call("rbind", list(fung_deg_all, bact_deg_all, cross_deg_all))
all_deg_dat$type <- factor(all_deg_dat$type, levels = c("Fungi", "Bacteria", "Interkingdom"))

full_deg_boxplot <- ggplot(all_deg_dat, aes(x=habitat, y=log_val, fill=type)) +
    #geom_point(position = position_jitterdodge(dodge.width=0.7), alpha=0.3) +
    geom_boxplot(key_glyph = "rect") +
    scale_fill_manual(values =  paletteer_d("vapoRwave::floralShoppe")[2:4]) +
    #scale_color_manual(values =  paletteer_d("vapoRwave::floralShoppe")[2:4]) +
    theme_classic() +
    theme(axis.text.x=element_text(colour="black",size=9)) +
    theme(text=element_text(colour="black",size=12)) +
    theme(axis.text.y=element_text(colour="black",size=9)) +
    theme(axis.title.y=element_blank()) +
    theme(axis.title.x=element_blank()) +
    scale_y_continuous(breaks = seq(0, 6, 2), limits = c(c(0, 6))) +
    labs(title = "Degree, log-transformed", fill="Kingdom") +
    #theme(legend.position = "none") +
    #g +
    guides(fill = "none")


## btwn cent boxplots
all_bc_dat <- do.call("rbind", list(fung_bc_all, bact_bc_all, cross_bc_all))
all_bc_dat$type <- factor(all_bc_dat$type, levels = c("Fungi", "Bacteria", "Interkingdom"))

full_bc_boxplot <- ggplot(all_bc_dat, aes(x=habitat, y=log_val, fill=type)) +
    #geom_point(position = position_jitterdodge(dodge.width=0.7), alpha=0.3) +
    geom_boxplot(key_glyph = "rect") +
    scale_fill_manual(values =  paletteer_d("vapoRwave::floralShoppe")[2:4]) +
    scale_color_manual(values =  paletteer_d("vapoRwave::floralShoppe")[2:4]) +
    theme_classic() +
    theme(axis.text.x=element_text(colour="black",size=9)) +
    theme(text=element_text(colour="black",size=12)) +
    theme(axis.text.y=element_text(colour="black",size=9)) +
    theme(axis.title.y=element_blank()) +
    theme(axis.title.x=element_blank()) +
    scale_y_continuous(breaks = seq(0, 13, 2), limits = c(c(0, 12.4))) +
    labs(title = "Betweenness Centrality, log-transformed", fill="Kingdom") +
    #theme(legend.position = "none") +
    #g
    guides(fill = "none")



library(patchwork)

library(cowplot)
my_legend <- get_legend(full_metric_plots[[1]])
library(ggpubr)
pp <- as_ggplot(my_legend)


boot_the_legend <- function(ggplot) {
  x = ggplot + theme(legend.position = "none")
  return(x)
}

full_metric_plots <- lapply(full_metric_plots, boot_the_legend)

conn_pl <- conn_pl + theme(legend.position = "none")

all_plots <- full_deg_boxplot + full_bc_boxplot + pp + full_metric_plots[[1]] + full_metric_plots[[2]] + full_metric_plots[[3]] 

ggsave("../figures/aug2024_revisions/metrics_bars/habitat_metrics_collected.pdf", width=15, height=9)
```



# Downsampled

```{r}
# import igraph objects
fung_igraphs <- readRDS("../intermediates/richtest_downsamp/fung_hab_downsamp721_igraphs.rds")
bact_igraphs <- readRDS("../intermediates/richtest_downsamp/bact_hab_downsamp721_igraphs.rds")
cross_igraphs <- readRDS("../intermediates/richtest_downsamp/cross_hab_downsamp721_igraphs.rds")

habs = c("Stream", "Terrestrial", "Marine")
```

### Metrics to Calculate

## ITS

```{r}
fung_nets <- fung_igraphs


## BETWEENNESS CENTRALITY
fung_bc <- lapply(fung_nets, betweenness)

## DEGREE
fung_deg <- lapply(fung_nets, degree)

## CONNECTANCE
connectance <- function(g) {
  ecount(g) / vcount(g)^2 
}


fung_conn <- sapply(fung_nets, connectance)


## EDGE TO NODE RATIO
fung_num_nodes <- sapply(fung_nets, gorder)
fung_num_edges <- sapply(fung_nets, gsize)

fung_eton_ratio <- fung_num_edges / fung_num_nodes

## MODULARITY
fung_netcarto <- readRDS("../intermediates/habitat/hab_downsamp721_netcarto.rds")[1:3]

fung_new_mod <- sapply(fung_netcarto, function(x) x[[2]])



metrics <- c("Modularity", "Connectance", "Edge to Node Ratio")

fung_metrics <- data.frame(modularity <- fung_new_mod,
                           connectance <- fung_conn,
                           eton_ratio <- fung_eton_ratio
                           )
names(fung_metrics) <- metrics
fung_metrics$habitat <- habs

fung_mt_long <- melt(fung_metrics, id.vars = "habitat", variable.name = "category", value.name = "value")
names(fung_mt_long) <- c("habitat", "metric", "value")
fung_mt_long$type <- "Fungi"


## node metric boxplots
fung_deg_dat <- list()
for (i in 1:3) {
  df <- data.frame(habitat = rep(habs[[i]], length(fung_deg[[i]])),
                   metric = rep("degree", length(fung_deg[[i]])),
                   value = fung_deg[[i]])
  
  fung_deg_dat[[i]] <- df
}

fung_deg_all <- do.call("rbind", fung_deg_dat)
fung_deg_all$log_val <- log(fung_deg_all$value + 1)
fung_deg_all$type <- "Fungi"


fung_bc_dat <- list()
for (i in 1:3) {
  df <- data.frame(habitat = rep(habs[[i]], length(fung_bc[[i]])),
                   metric = rep("betweenness", length(fung_bc[[i]])),
                   value = fung_bc[[i]])

  fung_bc_dat[[i]] <- df
}

fung_bc_all <- do.call("rbind", fung_bc_dat)
fung_bc_all$log_val <- log(fung_bc_all$value + 1)
fung_bc_all$type <- "Fungi"

```


## 16S

```{r}
bact_nets <- bact_igraphs


## BETWEENNESS CENTRALITY
bact_bc <- lapply(bact_nets, betweenness)

## DEGREE
bact_deg <- lapply(bact_nets, degree)

## CONNECTANCE
connectance <- function(g) {
  ecount(g) / vcount(g)^2 
}


bact_conn <- sapply(bact_nets, connectance)


## EDGE TO NODE RATIO
bact_num_nodes <- sapply(bact_nets, gorder)
bact_num_edges <- sapply(bact_nets, gsize)

bact_eton_ratio <- bact_num_edges / bact_num_nodes

## MODULARITY
bact_netcarto <- readRDS("../intermediates/habitat/hab_downsamp721_netcarto.rds")[4:6]

bact_new_mod <- sapply(bact_netcarto, function(x) x[[2]])



metrics <- c("Modularity", "Connectance", "Edge to Node Ratio")

bact_metrics <- data.frame(modularity <- bact_new_mod,
                           connectance <- bact_conn,
                           eton_ratio <- bact_eton_ratio
                           )
names(bact_metrics) <- metrics
bact_metrics$habitat <- habs

bact_mt_long <- melt(bact_metrics, id.vars = "habitat", variable.name = "category", value.name = "value")
names(bact_mt_long) <- c("habitat", "metric", "value")
bact_mt_long$type <- "Bacteria"


## node metric boxplots
bact_deg_dat <- list()
for (i in 1:3) {
  df <- data.frame(habitat = rep(habs[[i]], length(bact_deg[[i]])),
                   metric = rep("degree", length(bact_deg[[i]])),
                   value = bact_deg[[i]])
  
  bact_deg_dat[[i]] <- df
}

bact_deg_all <- do.call("rbind", bact_deg_dat)
bact_deg_all$log_val <- log(bact_deg_all$value + 1)
bact_deg_all$type <- "Bacteria"


bact_bc_dat <- list()
for (i in 1:3) {
  df <- data.frame(habitat = rep(habs[[i]], length(bact_bc[[i]])),
                   metric = rep("betweenness", length(bact_bc[[i]])),
                   value = bact_bc[[i]])

  bact_bc_dat[[i]] <- df
}

bact_bc_all <- do.call("rbind", bact_bc_dat)
bact_bc_all$log_val <- log(bact_bc_all$value + 1)
bact_bc_all$type <- "Bacteria"

```


## Cross

```{r}
cross_nets <- cross_igraphs


## BETWEENNESS CENTRALITY
cross_bc <- lapply(cross_nets, betweenness)

## DEGREE
cross_deg <- lapply(cross_nets, degree)

## CONNECTANCE
connectance <- function(g) {
  ecount(g) / vcount(g)^2 
}


cross_conn <- sapply(cross_nets, connectance)


## EDGE TO NODE RATIO
cross_num_nodes <- sapply(cross_nets, gorder)
cross_num_edges <- sapply(cross_nets, gsize)

cross_eton_ratio <- cross_num_edges / cross_num_nodes

## MODULARITY
cross_netcarto <- readRDS("../intermediates/habitat/hab_downsamp721_netcarto.rds")[7:9]

cross_new_mod <- sapply(cross_netcarto, function(x) x[[2]])



metrics <- c("Modularity", "Connectance", "Edge to Node Ratio")

cross_metrics <- data.frame(modularity <- cross_new_mod,
                           connectance <- cross_conn,
                           eton_ratio <- cross_eton_ratio
                           )
names(cross_metrics) <- metrics
cross_metrics$habitat <- habs

cross_mt_long <- melt(cross_metrics, id.vars = "habitat", variable.name = "category", value.name = "value")
names(cross_mt_long) <- c("habitat", "metric", "value")
cross_mt_long$type <- "Interkingdom"


## node metric boxplots
cross_deg_dat <- list()
for (i in 1:3) {
  df <- data.frame(habitat = rep(habs[[i]], length(cross_deg[[i]])),
                   metric = rep("degree", length(cross_deg[[i]])),
                   value = cross_deg[[i]])
  
  cross_deg_dat[[i]] <- df
}

cross_deg_all <- do.call("rbind", cross_deg_dat)
cross_deg_all$log_val <- log(cross_deg_all$value + 1)
cross_deg_all$type <- "Interkingdom"


cross_bc_dat <- list()
for (i in 1:3) {
  df <- data.frame(habitat = rep(habs[[i]], length(cross_bc[[i]])),
                   metric = rep("betweenness", length(cross_bc[[i]])),
                   value = cross_bc[[i]])

  cross_bc_dat[[i]] <- df
}

cross_bc_all <- do.call("rbind", cross_bc_dat)
cross_bc_all$log_val <- log(cross_bc_all$value + 1)
cross_bc_all$type <- "Interkingdom"

```


# Combine bars

```{r}
all_metric_data <- do.call("rbind", list(fung_mt_long, bact_mt_long, cross_mt_long))
all_metric_data$value <- as.numeric(all_metric_data$value)

all_metric_data$type <- factor(all_metric_data$type, levels = c("Fungi", "Bacteria", "Interkingdom"))


# consistent guides
g <- guides(fill = guide_legend(override.aes = list(fill =  paletteer_d("vapoRwave::floralShoppe")[2:4],
                                                    color = "black")))

unadjusted_full_metric_plots <- list()

for (i in 1:length(metrics)) {
  dat <- all_metric_data[all_metric_data$metric==metrics[i],]
  
  pl <- ggplot(dat, aes(x = habitat, y = value, fill = type)) +
    geom_bar(stat="identity", position="dodge", key_glyph = "rect") +
    scale_fill_manual(values =  paletteer_d("vapoRwave::floralShoppe")[2:4]) +
    theme_classic() +
    theme(axis.text.x=element_text(colour="black",size=9)) +
    theme(text=element_text(colour="black",size=12)) +
    theme(axis.text.y=element_text(colour="black",size=9)) +
    theme(axis.title.y=element_blank()) +
    theme(axis.title.x=element_blank()) +
    labs(title = metrics[i], fill = "Kingdom") +
    g
  
  unadjusted_full_metric_plots[[i]] <- pl
}

full_metric_plots <- unadjusted_full_metric_plots

# modularity
full_metric_plots[[1]] <- unadjusted_full_metric_plots[[1]] +
                          scale_y_continuous(breaks = seq(0, 0.7, 0.1),
                                             limits = c(c(0, 0.7)))

full_metric_plots[[2]] <- unadjusted_full_metric_plots[[2]] +
                          scale_y_continuous(breaks = seq(0, 0.008, 0.002),
                                             limits = c(c(0, 0.008)))


full_metric_plots[[3]] <- unadjusted_full_metric_plots[[3]] +
                          scale_y_continuous(breaks = seq(0, 6, 2),
                                             limits = c(c(0, 6)))



## degree boxplots
all_deg_dat <- do.call("rbind", list(fung_deg_all, bact_deg_all, cross_deg_all))
all_deg_dat$type <- factor(all_deg_dat$type, levels = c("Fungi", "Bacteria", "Interkingdom"))

full_deg_boxplot <- ggplot(all_deg_dat, aes(x=habitat, y=log_val, fill=type)) +
    #geom_point(position = position_jitterdodge(dodge.width=0.7), alpha=0.3) +
    geom_boxplot(key_glyph = "rect") +
    scale_fill_manual(values =  paletteer_d("vapoRwave::floralShoppe")[2:4]) +
    #scale_color_manual(values =  paletteer_d("vapoRwave::floralShoppe")[2:4]) +
    theme_classic() +
    theme(axis.text.x=element_text(colour="black",size=9)) +
    theme(text=element_text(colour="black",size=12)) +
    theme(axis.text.y=element_text(colour="black",size=9)) +
    theme(axis.title.y=element_blank()) +
    theme(axis.title.x=element_blank()) +
    #scale_y_continuous(breaks = seq(0, 6, 2), limits = c(c(0, 6))) +
    labs(title = "Degree, log-transformed", fill="Kingdom") +
    #theme(legend.position = "none") +
    #g +
    guides(fill = "none")


## btwn cent boxplots
all_bc_dat <- do.call("rbind", list(fung_bc_all, bact_bc_all, cross_bc_all))
all_bc_dat$type <- factor(all_bc_dat$type, levels = c("Fungi", "Bacteria", "Interkingdom"))

full_bc_boxplot <- ggplot(all_bc_dat, aes(x=habitat, y=log_val, fill=type)) +
    #geom_point(position = position_jitterdodge(dodge.width=0.7), alpha=0.3) +
    geom_boxplot(key_glyph = "rect") +
    scale_fill_manual(values =  paletteer_d("vapoRwave::floralShoppe")[2:4]) +
    scale_color_manual(values =  paletteer_d("vapoRwave::floralShoppe")[2:4]) +
    theme_classic() +
    theme(axis.text.x=element_text(colour="black",size=9)) +
    theme(text=element_text(colour="black",size=12)) +
    theme(axis.text.y=element_text(colour="black",size=9)) +
    theme(axis.title.y=element_blank()) +
    theme(axis.title.x=element_blank()) +
    #scale_y_continuous(breaks = seq(0, 13, 2), limits = c(c(0, 12.4))) +
    labs(title = "Betweenness Centrality, log-transformed", fill="Kingdom") +
    #theme(legend.position = "none") +
    #g
    guides(fill = "none")



library(patchwork)

library(cowplot)
my_legend <- get_legend(full_metric_plots[[1]])
library(ggpubr)
pp <- as_ggplot(my_legend)


boot_the_legend <- function(ggplot) {
  x = ggplot + theme(legend.position = "none")
  return(x)
}

full_metric_plots <- lapply(full_metric_plots, boot_the_legend)


all_plots <- full_deg_boxplot + full_bc_boxplot + pp + full_metric_plots[[1]] + full_metric_plots[[2]] + full_metric_plots[[3]] 

ggsave("../figures/aug2024_revisions/metrics_bars/downsamp_habitat_metrics_collected.pdf", width=15, height=9)
```
