==> habitat_networks/05b_newdownsamp_habitat_robustness.Rmd <==
---
title: "05_habitat_robustness"
output: html_document
date: "2023-06-06"
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

```{r, message=F}
library(dplyr)
library(igraph)
library(ggnetwork)
library(intergraph)
library(brainGraph)
library(patchwork)
library(network)
library(SpiecEasi)

library(dplyr)
library(tibble)

options(bitmapType='cairo')
```

```{r}
# import igraph objects
# fung_igraphs <- readRDS("../intermediates/habitat/fung_habitat_igraphs.rds")
# bact_igraphs <- readRDS("../intermediates/habitat/bact_habitat_igraphs.rds")
# cross_igraphs <- readRDS("../intermediates/habitat/cross_habitat_igraphs.rds")
# 
# # metadata
# fung_meta <- readRDS("../intermediates/global/fully_filtered_p20_fungal_otu_metadata_matched_up.rds")
# bact_meta <- readRDS("../intermediates/global/fully_filtered_p20_bact_otu_metadata_matched_up.rds")
# 
# habs <- unique(fung_meta$habitat)
# 
# 
# source("~/hynson_koastore/kaciekaj/waimea_current/misc/new_knockout_functions.R")
```

# Marine

```{r}
fung_bc_rob = lapply(fung_igraphs, 
                     function(x) myrobustness_bc(x, kingdom = "Fungi"))

fung_invbc_rob = lapply(fung_igraphs, 
                        function(x) myrobustness_inversebc(x, kingdom = "Fungi"))

fung_random_rob = lapply(fung_igraphs,
                         function(x) alt_robustness(x, type = "vertex", "random", kingdom = "Fungi"))

saveRDS(list(fung_bc_rob, fung_invbc_rob, fung_random_rob),
        "~/hynson_koastore/kaciekaj/waimea_current/intermediates/new_knockouts/fung_habitat_knockouts.rds")


bact_bc_rob = myrobustness_bc(filt_bact_net)

bact_invbc_rob = myrobustness_inversebc(filt_bact_net)

bact_random_rob = alt_robustness(filt_bact_net, type = "vertex", "random")



cross_bc_rob = myrobustness_bc(filt_cross_net)

cross_invbc_rob = myrobustness_inversebc(filt_cross_net)

cross_random_rob = alt_robustness(filt_cross_net, type = "vertex", "random")

```

```{r}
# stream, terrestrial, marine

# robustness results from slurm
# fung_robustness <- readRDS("../intermediates/new_knockouts/fung_habitat_knockouts.rds")
# bact_robustness <- readRDS("../intermediates/new_knockouts/bact_habitat_knockouts.rds")
# cross_robustness <- readRDS("../intermediates/new_knockouts/cross_habitat_knockouts.rds")

fung_rob <- readRDS("../intermediates/new_knockouts/new_downsamp719_all_network_bc_knockouts.rds")[4:6]
bact_rob <- readRDS("../intermediates/new_knockouts/new_downsamp719_all_network_bc_knockouts.rds")[7:9]
cross_rob <- readRDS("../intermediates/new_knockouts/new_downsamp719_all_network_bc_knockouts.rds")[10:12]

prep_tbl <- function(list_of_robustness_tables,  kingdom) {
  dat <- do.call("rbind", list_of_robustness_tables)
  dat$kingdom = c(rep(kingdom, nrow(dat)))
  
  return(dat)
}

fung_robustness <- prep_tbl(fung_rob, "Fungi")
bact_robustness <- prep_tbl(bact_rob, "Bacteria")
cross_robustness <- prep_tbl(cross_rob, "Interkingdom")


fung_robustness$habitat = c(rep("Stream", 720), 
                  rep("Terrestrial", 720), 
                  rep("Marine", 720))

bact_robustness$habitat = c(rep("Stream", 720), 
                  rep("Terrestrial", 720), 
                  rep("Marine", 720))

cross_robustness$habitat = c(rep("Stream", 720), 
                  rep("Terrestrial", 720), 
                  rep("Marine", 720))

```

```{r}
# fung_bc_rob <- fung_robustness[[1]]
# fung_invbc_rob <- fung_robustness[[2]]
# fung_random_rob <- fung_robustness[[3]]
# 
# bact_bc_rob <- bact_robustness[[1]]
# bact_invbc_rob <- bact_robustness[[2]]
# bact_random_rob <- bact_robustness[[3]]
# 
# cross_bc_rob <- cross_robustness[[1]]
# cross_invbc_rob <- cross_robustness[[2]]
# cross_random_rob <- cross_robustness[[3]]


# all_stream <- rbind(fung_bc_rob[[1]], fung_invbc_rob[[1]], fung_random_rob[[1]],
#                         bact_bc_rob[[1]], bact_invbc_rob[[1]], bact_random_rob[[1]],
#                         cross_bc_rob[[1]], cross_invbc_rob[[1]], cross_random_rob[[1]]
#                         )

all_stream <- rbind(fung_robustness[fung_robustness$habitat=="Stream",],
                    bact_robustness[bact_robustness$habitat=="Stream",],
                    cross_robustness[cross_robustness$habitat=="Stream",])
#all_stream$category <- paste(all_stream$type, all_stream$kingdom, sep="_")
#all_stream$habitat <- rep("Stream", nrow(all_stream))
#all_stream$category2 <- paste(all_stream$type, all_stream$habitat, sep = "_")


# all_terrestrial <- rbind(fung_bc_rob[[2]], fung_invbc_rob[[2]], fung_random_rob[[2]],
#                         bact_bc_rob[[2]], bact_invbc_rob[[2]], bact_random_rob[[2]],
#                         cross_bc_rob[[2]], cross_invbc_rob[[2]], cross_random_rob[[2]]
#                         )
# all_terrestrial$category <- paste(all_terrestrial$type, all_terrestrial$kingdom, sep="_")
# all_terrestrial$habitat <- rep("Terrestrial", nrow(all_terrestrial))
# all_terrestrial$category2 <- paste(all_terrestrial$type, all_terrestrial$habitat, sep = "_")

all_terr <- rbind(fung_robustness[fung_robustness$habitat=="Terrestrial",],
                    bact_robustness[bact_robustness$habitat=="Terrestrial",],
                    cross_robustness[cross_robustness$habitat=="Terrestrial",])

# all_marine <- rbind(fung_bc_rob[[3]], fung_invbc_rob[[3]], fung_random_rob[[3]],
#                         bact_bc_rob[[3]], bact_invbc_rob[[3]], bact_random_rob[[3]],
#                         cross_bc_rob[[3]], cross_invbc_rob[[3]], cross_random_rob[[3]]
#                         )
# all_marine$category <- paste(all_marine$type, all_marine$kingdom, sep="_")
# all_marine$habitat <- rep("Marine", nrow(all_marine))
# all_marine$category2 <- paste(all_marine$type, all_marine$habitat, sep = "_")

all_marine <- rbind(fung_robustness[fung_robustness$habitat=="Marine",],
                    bact_robustness[bact_robustness$habitat=="Marine",],
                    cross_robustness[cross_robustness$habitat=="Marine",])

```


```{r}

library(paletteer)
kingdom_palette = paletteer_d("fishualize::Histiophryne_psychedelica")


stream_bc_plot <- ggplot(all_stream, aes(x=removed.pct, y=comp.pct, color=kingdom, group = kingdom, linetype = type)) +
  geom_line(linewidth = 1.5) +
  labs(x="Fraction of nodes removed", y="Fractional size of largest connected component",
color = "Category", shape = "Kingdom", title = "Stream", linetype = "Type") +
  scale_color_manual(values = kingdom_palette) +
  theme(legend.key = element_rect(fill="white")) +
  theme(panel.border = element_blank(), panel.grid.major = element_blank(),
        panel.grid.minor = element_blank(), axis.line = element_line(colour = "black"),
        panel.background=element_blank(),
        plot.title=element_text(size=16),
        axis.title = element_text(size=16),
        axis.text = element_text(color = "black", size= 12),
        title=element_text(size=12)) +
  theme(legend.text = element_text(size=12)) +
   theme(legend.title = element_text(size=14)) +
   guides(color = "none", linetype = "none") +
    scale_linetype_manual(values = c("btwncent" = "solid", "inverse_btwncent" = "dashed", 
                                   "random" = "dotted")) +
  theme(axis.title.y = element_text(margin = margin(t = 0, r = 10, b = 0, l = 0))) +
  theme(axis.title.x = element_text(margin = margin(t = 10, r = 0, b = 0, l = 0))) +
  scale_x_continuous(limits = c(0,1), expand = c(0, 0)) +
  scale_y_continuous(limits = c(0,1), expand = c(0, 0)) 


terr_bc_plot <- ggplot(all_terr, aes(x=removed.pct, y=comp.pct, color=kingdom, group = kingdom, linetype = type)) +
  geom_line(linewidth = 1.5) +
  labs(x="Fraction of nodes removed", y="Fractional size of largest connected component",
color = "Category", shape = "Kingdom", title = "Terrestrial", linetype = "Type") +
   scale_color_manual(values = kingdom_palette) +
  theme(legend.key = element_rect(fill="white")) +
  theme(panel.border = element_blank(), panel.grid.major = element_blank(),
        panel.grid.minor = element_blank(), axis.line = element_line(colour = "black"),
        panel.background=element_blank(),
        plot.title=element_text(size=16),
        axis.title = element_text(size=16),
        axis.text = element_text(color = "black", size= 12),
        title=element_text(size=12)) +
  theme(legend.text = element_text(size=12)) +
   theme(legend.title = element_text(size=14)) +
   guides(color = "none", linetype = "none") +
    scale_linetype_manual(values = c("btwncent" = "solid", "inverse_btwncent" = "dashed", 
                                   "random" = "dotted")) +
  theme(axis.title.y = element_text(margin = margin(t = 0, r = 10, b = 0, l = 0))) +
  theme(axis.title.x = element_text(margin = margin(t = 10, r = 0, b = 0, l = 0))) +
  scale_x_continuous(limits = c(0,1), expand = c(0, 0)) +
  scale_y_continuous(limits = c(0,1), expand = c(0, 0)) 


marine_bc_plot <- ggplot(all_marine, aes(x=removed.pct, y=comp.pct, color=kingdom, group = kingdom, linetype = type)) +
  geom_line(linewidth = 1.5) +
  labs(x="Fraction of nodes removed", y="Fractional size of largest connected component",
color = "Category", shape = "Kingdom", title = "Marine", linetype = "Type") +
   scale_color_manual(values = kingdom_palette) +
  theme(legend.key = element_rect(fill="white")) +
  theme(panel.border = element_blank(), panel.grid.major = element_blank(),
        panel.grid.minor = element_blank(), axis.line = element_line(colour = "black"),
        panel.background=element_blank(),
        plot.title=element_text(size=16),
        axis.title = element_text(size=16),
        axis.text = element_text(color = "black", size= 12),
        title=element_text(size=12)) +
  theme(legend.text = element_text(size=12)) +
   theme(legend.title = element_text(size=14)) +
   #guides(color = "none") +
    scale_linetype_manual(values = c("btwncent" = "solid", "inverse_btwncent" = "dashed", 
                                   "random" = "dotted")) +
  theme(axis.title.y = element_text(margin = margin(t = 0, r = 10, b = 0, l = 0))) +
  theme(axis.title.x = element_text(margin = margin(t = 10, r = 0, b = 0, l = 0))) +
  scale_x_continuous(limits = c(0,1), expand = c(0, 0)) +
  scale_y_continuous(limits = c(0,1), expand = c(0, 0)) 


library(patchwork)

hab_plots <- stream_bc_plot + terr_bc_plot + marine_bc_plot


#ggsave("../figures/aug2024_revisions/habitat_robust_by_habitat.pdf", hab_plots, height = 6, width = 19)
#ggsave("../figures/aug2024_revisions/habitat_robust_by_habitat.png", hab_plots,  height = 6, width = 19)
```

# Fig 2 - bringing over code from watershed networks

```{r}
all_robustness <- readRDS("../intermediates/new_knockouts/whole_watershed_bc_knockouts.rds")

all_robustness$category <- paste(all_robustness$type, all_robustness$kingdom, sep="_")

library(paletteer)
#habitat_palette = paletteer_d("waRhol::camo_87_5")[c(1,2,4)]
kingdom_palette = paletteer_d("fishualize::Histiophryne_psychedelica")


bc_plot <- ggplot(all_robustness, aes(x=removed.pct, y=comp.pct, color=kingdom, group = category, linetype = type)) +
  geom_line(linewidth = 1.5) +
  labs(x="Fraction of nodes removed", y="Fractional size of largest connected component",
color = "Category", shape = "Kingdom", title = "Whole Watershed", linetype = "Type") +
  scale_color_manual(values = kingdom_palette) +
  theme(legend.key = element_rect(fill="white")) +
  theme(panel.border = element_blank(), panel.grid.major = element_blank(),
        panel.grid.minor = element_blank(), axis.line = element_line(colour = "black"),
        panel.background=element_blank(),
        plot.title=element_text(size=16),
        axis.title = element_text(size=16),
        axis.text = element_text(size = 12, color = "black"),
        title=element_text(size=12)) +
  theme(legend.text = element_text(size=12)) +
   theme(legend.title = element_text(size=14)) +
  guides(color = "none", linetype = "none") +
  scale_linetype_manual(values = c("btwncent" = "solid", "inverse_btwncent" = "dashed", 
                                   "random" = "dotted")) +
  theme(axis.title.y = element_text(margin = margin(t = 0, r = 10, b = 0, l = 0))) +
  theme(axis.title.x = element_text(margin = margin(t = 10, r = 0, b = 0, l = 0))) +
  scale_x_continuous(limits = c(0,1), expand = c(0, 0)) +
  scale_y_continuous(limits = c(0,1), expand = c(0, 0)) 


all <- bc_plot + stream_bc_plot + terr_bc_plot + marine_bc_plot + plot_layout(ncol = 2)

ggsave("../figures/aug2024_revisions/watershed_and_habitat_curves_by_kingdom.png", all, height = 10, width = 12)
ggsave("../figures/aug2024_revisions/watershed_and_habitat_curves_by_kingdom.pdf", all, height = 10, width = 12)
```


## By kingdom

```{r}
all_fung_dat <- rbind(all_stream[all_stream$kingdom=="Fungi",],
                      all_terrestrial[all_terrestrial$kingdom=="Fungi",],
                      all_marine[all_marine$kingdom=="Fungi",])

all_bact_dat <- rbind(all_stream[all_stream$kingdom=="Bacteria",],
                      all_terrestrial[all_terrestrial$kingdom=="Bacteria",],
                      all_marine[all_marine$kingdom=="Bacteria",])

all_cross_dat <- rbind(all_stream[all_stream$kingdom=="Interkingdom",],
                      all_terrestrial[all_terrestrial$kingdom=="Interkingdom",],
                      all_marine[all_marine$kingdom=="Interkingdom",])
```


```{r}
habitat_palette <- paletteer_d("ltc::ploen")[c(1,2,3)]

fung_bc_plot <- ggplot(fung_robustness, aes(x=removed.pct, y=comp.pct, color=habitat, group = habitat, linetype = type)) +
  geom_line(linewidth = 1.5) +
  labs(x="Fraction of nodes removed", y="Fractional size of largest connected component",
color = "Category", shape = "Kingdom", title = "Fungi", linetype = "Type") +
  scale_color_manual(values = habitat_palette) +
  theme(legend.key = element_rect(fill="white")) +
  theme(panel.border = element_blank(), panel.grid.major = element_blank(),
        panel.grid.minor = element_blank(), axis.line = element_line(colour = "black"),
        panel.background=element_blank(),
        plot.title=element_text(size=16),
        axis.title = element_text(size=16),
        axis.text = element_text(color = "black", size = 12),
        title=element_text(size=12)) +
  theme(legend.text = element_text(size=12)) +
   theme(legend.title = element_text(size=14)) +
  guides(color = "none", linetype = "none") +
    scale_linetype_manual(values = c("btwncent" = "solid", "inverse_btwncent" = "dashed", 
                                   "random" = "dotted")) +
  theme(axis.title.y = element_text(margin = margin(t = 0, r = 10, b = 0, l = 0))) +
  theme(axis.title.x = element_text(margin = margin(t = 10, r = 0, b = 0, l = 0))) +
  scale_x_continuous(limits = c(0,1), expand = c(0, 0)) +
  scale_y_continuous(limits = c(0,1), expand = c(0, 0))



bact_bc_plot <- ggplot(bact_robustness, aes(x=removed.pct, y=comp.pct, color=habitat, group = habitat, linetype = type)) +
  geom_line(linewidth = 1.5) +
  labs(x="Fraction of nodes removed", y="Fractional size of largest connected component",
color = "Category", shape = "Kingdom", title = "Bacteria", linetype = "Type") +
  theme(legend.key = element_rect(fill="white")) +
  scale_color_manual(values = habitat_palette) +
  theme(panel.border = element_blank(), panel.grid.major = element_blank(),
        panel.grid.minor = element_blank(), axis.line = element_line(colour = "black"),
        panel.background=element_blank(),
        plot.title=element_text(size=16),
        axis.title = element_text(size=16),
        axis.text = element_text(color = "black", size = 12),
        title=element_text(size=12)) +
  theme(legend.text = element_text(size=12)) +
   theme(legend.title = element_text(size=14)) +
  guides(color = "none", linetype = "none") +
    scale_linetype_manual(values = c("btwncent" = "solid", "inverse_btwncent" = "dashed", 
                                   "random" = "dotted")) +
  theme(axis.title.y = element_text(margin = margin(t = 0, r = 10, b = 0, l = 0))) +
  theme(axis.title.x = element_text(margin = margin(t = 10, r = 0, b = 0, l = 0))) +
  scale_x_continuous(limits = c(0,1), expand = c(0, 0)) +
  scale_y_continuous(limits = c(0,1), expand = c(0, 0))


cross_bc_plot <- ggplot(cross_robustness, aes(x=removed.pct, y=comp.pct, color=habitat, group = habitat, linetype = type)) +
  geom_line(linewidth = 1.5) +
  labs(x="Fraction of nodes removed", y="Fractional size of largest connected component",
color = "Category", shape = "Kingdom", title = "Interkingdom", linetype = "Type") +
  scale_color_manual(values = habitat_palette) +
  theme(legend.key = element_rect(fill="white")) +
  theme(panel.border = element_blank(), panel.grid.major = element_blank(),
        panel.grid.minor = element_blank(), axis.line = element_line(colour = "black"),
        panel.background=element_blank(),
        plot.title=element_text(size=16),
        axis.title = element_text(size=16),
        axis.text = element_text(color = "black", size = 12),
        title=element_text(size=12)) +
  theme(legend.text = element_text(size=12)) +
   theme(legend.title = element_text(size=14)) +
  #guides(color = "none") +
    scale_linetype_manual(values = c("btwncent" = "solid", "inverse_btwncent" = "dashed", 
                                   "random" = "dotted")) +
  theme(axis.title.y = element_text(margin = margin(t = 0, r = 10, b = 0, l = 0))) +
  theme(axis.title.x = element_text(margin = margin(t = 10, r = 0, b = 0, l = 0))) +
  scale_x_continuous(limits = c(0,1), expand = c(0, 0)) +
  scale_y_continuous(limits = c(0,1), expand = c(0, 0))


bb = fung_bc_plot + bact_bc_plot + cross_bc_plot

#ggsave("../figures/aug2024_revisions/habitat_robust_by_kingdom.pdf", bb, height = 5.5, width = 18)
#ggsave("../figures/aug2024_revisions/habitat_robust_by_kingdom.png", bb, height = 5.5, width = 18)
```
