==> habitat_networks/06b_phylo_dist_habitat.Rmd <==
---
title: "05_phylo_dist_habitat"
output: html_document
date: "2023-06-22"
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

```{r, messsage=FALSE}
library(ape)
library(picante)
library(plyr)
library(ggplot2)

library(ggsignif)

options(bitmapType='cairo')
```

```{r}
bact_data <- readRDS("../intermediates/habitat/bact_downsampled_otu_tables_by_hab.rds")

full_bact_meta <- readRDS("../intermediates/global/fully_filtered_p20_bact_otu_metadata_matched_up.rds")

full_bact_physeq <- readRDS("../intermediates/gradient/bact_physeq_with_tree.rds")
sample_names(full_bact_physeq) <- full_bact_meta$sample_id[match(sample_names(full_bact_physeq),
                                                                full_bact_meta$x_seq_id)]


bact_tree <- phy_tree(full_bact_physeq)

habs <- unique(full_bact_meta$habitat)

# remove dummy row
for (i in 1:3) {
  dat <- bact_data[[i]]
  dat <- dat[rownames(dat)!="dummy",]
  bact_data[[i]] <- dat
}
```

```{r}
# rows should correspond to samples, meaning plot
# columns should be OTUS

# can I still use the whole tree though? I think 
dat_by_sample <- lapply(bact_data, function(x) as.data.frame(t(x)))

comm_otu_table <- do.call("rbind.fill", dat_by_sample)
comm_otu_table[is.na(comm_otu_table)] <- 0

my_samp_names <- lapply(dat_by_sample, function(x) rownames(x))

rownames(comm_otu_table) <- do.call("c", my_samp_names)


#pd_result <- picante::pd(comm_otu_table, bact_tree, include.root = FALSE)


phydist <- cophenetic(bact_tree)

stuff_for_picante <- list(abun = comm_otu_table,
                          dist = phydist)

saveRDS(stuff_for_picante, "../intermediates/habitat/inputs_for_picante_by_sample.rds")

```


## MPD by Sample

```{r}
# Positive SES values (mpd.obs.z > 0) and high quantiles (mpd.obs.p > 0.95) in-
# dicate phylogenetic evenness, or a greater phylogenetic distance among co-occurring
# species than expected. Negative SES values and low quantiles (mpd.obs.p < 0.05)
# indicate phylogenetic clustering, or small phylogenetic distances among co-occurring
# species than expected
samp_dat <- readRDS("../intermediates/habitat/picante_by_sample_habitat_results.rds")
samp_dat$sample <- rownames(samp_dat)

samp_dat$habitat <- full_bact_meta$habitat[match(samp_dat$sample, full_bact_meta$sample_id)]


library(wesanderson)
mpd_bar <- ggplot(samp_dat, aes(x=habitat, y=mpd.obs.z)) +
    geom_jitter(color=wes_palette("GrandBudapest2")[2], size=0.8, alpha=0.7, width=0.2) +
    geom_boxplot(width=0.5,fill = wes_palette("GrandBudapest2")[2]) +
    scale_y_continuous(breaks = seq(-3, 6, 3), limits = c(c(-3, 6))) +
    theme(axis.text.x = element_text(size=12, colour="black"), 
          panel.grid.major=element_blank(),
          panel.grid.minor=element_blank(),
          panel.background=element_blank(),
          axis.line = element_line(colour="black"),
          axis.text.y = element_text(size=12, colour="black"),
          axis.title = element_text(size=16),
          plot.title = element_text(size=18),
          legend.position = "none") +
  theme(axis.title.y = element_text(margin = margin(t = 0, r = 10, b = 0, l = 0))) +
  theme(axis.title.x = element_text(margin = margin(t = 10, r = 0, b = 0, l = 0))) +
  labs(title= "Mean Phylogenetic Distances by Sample", y = "Standardized Effect Score", x="Habitat") +
  geom_signif(comparisons = list(
    c("Marine", "Riverine"),
    c("Marine", "Terrestrial"),
    c("Riverine", "Terrestrial")
  ),
    map_signif_level = TRUE,
    y_position = c(3.3, 3.9, 4.8))

ggsave("../figures/finals/habitat/mean_phylodist_by_habitat_and_sample.pdf", width=8, height=7)


test_lm <- lm(mpd.obs.z ~ habitat, data = samp_dat)
test <- aov(mpd.obs.z ~ habitat, data = samp_dat)
tuk <- TukeyHSD(test)

tuk_df <- as.data.frame(tuk$habitat)

tuk_sig <- tuk_df[tuk_df$`p adj`<0.05,]

write.csv(tuk_df, "../outputs/habitat_mpd_tukey.csv")
```
